requirements.txt
asgiref==3.5.2
autopep8==1.6.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorclass==2.2.2
cryptography==38.0.3
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.16
django-allauth==0.51.0
django-environ==0.8.1
django-import-export==3.0.1
django-sslserver==0.22
docopt==0.6.2
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.4
MarkupPy==1.14
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
Pillow==9.3.0
pip-upgrader==1.4.15
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
tablib==3.2.1
terminaltables==3.1.10
toml==0.10.2
urllib3==1.26.12
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
